# argus-backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Argus AI — Event Collector API
  version: 2.1.0
  description: |
    REST API for the Argus AI proctoring platform.

    ## Authentication

    All endpoints (except `POST /api/v1/auth/login`) require a Bearer JWT in the
    `Authorization` header:

    ```
    Authorization: Bearer <token>
    ```

    Tokens are obtained via the login endpoint and expire after 24 hours.

    ## Rate Limits

    | Scope       | Limit             |
    |-------------|-------------------|
    | Global      | 5 000 req/s       |
    | Per-IP      | 100 req/s         |
    | Login       | 5 req/s (Nginx)   |

    Rate-limit responses return HTTP 429 with a `Retry-After` header.

  contact:
    name: Argus AI Engineering
    url: https://argus.ai
servers:
  - url: https://api.argus.ai
    description: Production
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/evidence/sessions/{sessionId}:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    get:
      tags: [Evidence]
      summary: List evidence fragments for session
      operationId: listEvidenceFragments
      responses:
        '200':
          description: Evidence fragment list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/evidence/fragments/{fragmentId}/url:
    parameters:
      - $ref: '#/components/parameters/FragmentId'
    get:
      tags: [Evidence]
      summary: Get presigned download URL
      description: "Returns a time-limited presigned S3 URL (default TTL: 5 minutes)."
      operationId: getEvidencePresignedURL
      responses:
        '200':
          description: Presigned URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresignedURLResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/evidence/fragments/{fragmentId}/verify:
    parameters:
      - $ref: '#/components/parameters/FragmentId'
    get:
      tags: [Evidence]
      summary: Verify fragment integrity
      description: Recomputes the SHA-256 hash of the S3 object and compares to the stored hash.
      operationId: verifyEvidenceFragment
      responses:
        '200':
          description: Integrity check result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyIntegrityResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Export ────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── Parameters ─────────────────────────────────────────────────────────────
  parameters:
...
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
      description: Proctoring session identifier
    FragmentId:
      name: fragmentId
      in: path
      required: true
      schema:
        type: string
      description: Evidence fragment identifier
...
  responses:
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "unauthorized"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "not found"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    EvidenceFragment:
      type: object
      properties:
        fragmentId:
          type: string
        sessionId:
          type: string
        eventId:
          type: string
        orgId:
          type: string
        examId:
          type: string
        studentId:
          type: string
        sha256Hash:
          type: string
          description: SHA-256 hex digest of the stored binary object
        uri:
          type: string
          description: Internal S3 URI (not directly accessible — use presigned URL endpoint)
        sizeBytes:
          type: integer
          format: int64
        contentType:
          type: string
        durationSec:
          type: number
          format: float
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        uploadedAt:
          type: string
          format: date-time
    EvidenceListResponse:
      type: object
      properties:
        sessionId:
          type: string
        fragments:
          type: array
          items:
            $ref: '#/components/schemas/EvidenceFragment'
        total:
          type: integer
        generatedAt:
          type: string
          format: date-time
    PresignedURLResponse:
      type: object
      properties:
        fragmentId:
          type: string
        url:
          type: string
          description: "Presigned HTTPS URL (default TTL: 5 minutes)"
        expiresAt:
          type: string
          format: date-time
        ttlSeconds:
          type: integer
    VerifyIntegrityResponse:
      type: object
      properties:
        fragmentId:
          type: string
        valid:
          type: boolean
        expectedHash:
          type: string
        message:
          type: string

    # ── Export ──────────────────────────────────────────────────────────────
...
security:
  - BearerAuth: []